Fix logging verbosity configuration bug
While our unit tests showed that the log configuration worked as intended, the more verbose loglevel messages were not shown on the commandline. We need to configure the root log handler to output verbose messages as well.

It's very important not to configure the root logger's level, as this has an influence on all other loggers running in all libraries etc., causing a lot of log noise on the CLI.

Thanks @TobiCode for discovering this bug.