Add integration test for --log-file
Quite some research and fiddling with Java and its annoyances involved, but it appears to work.

Test might be extended to do some more than "does the file exist" and "does it look like (broken) XML", but for that, we first need to decide whether to keep this log format.