Make sure parent path can be calculated by cf-control
Apparently, when passing a filename only, the platform-specific implementation in Paths isn't smart enough to assume e.g., a relative path like . as parent, and just returns a null pointer for paths like filename.yml. Since we're interested in the absolute path anyway, we can just fetch that before trying to query the parent path, and fix the bug that way.

In collaboration with @felix-oq