Avoid errors in apply while removing service
This should've been fixed before, but somehow the fix didn't make it into the codebase. This commit solves issues when a service is removed by the ServiceOperations (which unbinds them from the apps first). When the ApplicationsOperations then tries to unbind the service (because it doesn't exist any more), this operation can fail. This is a regular use case, and we only want to warn the user about this.